+++ spec/lib/aixm/document_spec.rb
@@ -243,6 +243,24 @@ describe AIXM::Document do
             <uomDist>M</uomDist>
             <txtRmk>forth remarks</txtRmk>
           </Rdd>
+          <Rls>
+            <RlsUid>
+              <RdnUid>
+                <RwyUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>16L/34R</txtDesig>
+                </RwyUid>
+                <txtDesig>16L</txtDesig>
+              </RdnUid>
+              <codePsn>AIM</codePsn>
+            </RlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Rls>
           <Rdn>
             <RdnUid>
               <RwyUid>
@@ -280,6 +298,24 @@ describe AIXM::Document do
             <uomDist>M</uomDist>
             <txtRmk>back remarks</txtRmk>
           </Rdd>
+          <Rls>
+            <RlsUid>
+              <RdnUid>
+                <RwyUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>16L/34R</txtDesig>
+                </RwyUid>
+                <txtDesig>34R</txtDesig>
+              </RdnUid>
+              <codePsn>AIM</codePsn>
+            </RlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Rls>
           <Fto>
             <FtoUid>
               <AhpUid>
@@ -324,6 +360,24 @@ describe AIXM::Document do
             <valMagBrg>356</valMagBrg>
             <txtRmk>Avoid flight over residental area</txtRmk>
           </Fdn>
+          <Fls>
+            <FlsUid>
+              <FdnUid>
+                <FtoUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>H1</txtDesig>
+                </FtoUid>
+                <txtDesig>35</txtDesig>
+              </FdnUid>
+              <codePsn>AIM</codePsn>
+            </FlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Fls>
           <Tla>
             <TlaUid>
               <AhpUid>
@@ -365,6 +419,21 @@ describe AIXM::Document do
             <codeSts>OTHER</codeSts>
             <txtRmk>Authorizaton by AD operator required</txtRmk>
           </Tla>
+          <Tls>
+            <TlsUid>
+              <TlaUid>
+                <AhpUid>
+                  <codeId>LFNT</codeId>
+                </AhpUid>
+                <txtDesig>H1</txtDesig>
+              </TlaUid>
+              <codePsn>AIM</codePsn>
+            </TlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Tls>
           <Ahu>
             <AhuUid>
               <AhpUid>
@@ -989,6 +1058,24 @@ describe AIXM::Document do
             <uomDist>M</uomDist>
             <txtRmk>forth remarks</txtRmk>
           </Rdd>
+          <Rls>
+            <RlsUid>
+              <RdnUid>
+                <RwyUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>16L/34R</txtDesig>
+                </RwyUid>
+                <txtDesig>16L</txtDesig>
+              </RdnUid>
+              <codePsn>AIM</codePsn>
+            </RlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Rls>
           <Rdn>
             <RdnUid>
               <RwyUid>
@@ -1026,6 +1113,24 @@ describe AIXM::Document do
             <uomDist>M</uomDist>
             <txtRmk>back remarks</txtRmk>
           </Rdd>
+          <Rls>
+            <RlsUid>
+              <RdnUid>
+                <RwyUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>16L/34R</txtDesig>
+                </RwyUid>
+                <txtDesig>34R</txtDesig>
+              </RdnUid>
+              <codePsn>AIM</codePsn>
+            </RlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Rls>
           <Fto>
             <FtoUid>
               <AhpUid>
@@ -1070,6 +1175,24 @@ describe AIXM::Document do
             <valMagBrg>356</valMagBrg>
             <txtRmk>Avoid flight over residental area</txtRmk>
           </Fdn>
+          <Fls>
+            <FlsUid>
+              <FdnUid>
+                <FtoUid>
+                  <AhpUid>
+                    <codeId>LFNT</codeId>
+                  </AhpUid>
+                  <txtDesig>H1</txtDesig>
+                </FtoUid>
+                <txtDesig>35</txtDesig>
+              </FdnUid>
+              <codePsn>AIM</codePsn>
+            </FlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Fls>
           <Tla>
             <TlaUid>
               <AhpUid>
@@ -1111,6 +1234,21 @@ describe AIXM::Document do
             <codeSts>OTHER</codeSts>
             <txtRmk>Authorizaton by AD operator required</txtRmk>
           </Tla>
+          <Tls>
+            <TlsUid>
+              <TlaUid>
+                <AhpUid>
+                  <codeId>LFNT</codeId>
+                </AhpUid>
+                <txtDesig>H1</txtDesig>
+              </TlaUid>
+              <codePsn>AIM</codePsn>
+            </TlsUid>
+            <txtDescr>omnidirectional</txtDescr>
+            <codeIntst>LIM</codeIntst>
+            <codeColour>GRN</codeColour>
+            <txtRmk>lighting remarks</txtRmk>
+          </Tls>
           <Ahu>
             <AhuUid>
               <AhpUid>